requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
langdetect==1.0.9
openai==1.3.0
gspread==5.12.0
oauth2client==4.1.3
python-multipart==0.0.6

